rename: 포트폴리오 작성 관련 컴포넌트 폴더명 변경
