+++ src/pages/TestPortfolioPage.tsx
@@ -1,4 +1,4 @@
-import CreatePortfolio from '@src/components/portfolio/CreatePortfolio';
+import CreatePortfolio from '@src/components/createPortfolio/CreatePortfolio';
 
 const TestPortfolioPage = () => {
   return <CreatePortfolio />;
